Function bulletproofs::inner_product_proof::prove_inner_product_with_scalars

pub fn prove_inner_product_with_scalars<C: Curve>(
    transcript: &mut RandomOracle, 
    G_slice: &[C], 
    H_slice: &[C], 
    H_prime_scalars: &[C::Scalar], 
    Q: &C, 
    a_slice: &[C::Scalar], 
    b_slice: &[C::Scalar]
) -> Option<InnerProductProof<C>>

This function computes an inner product proof, which is a proof of knowledge that the prover knows vectors a and b such that P’=<a,G>+<b,H’>+<a,b>Q, but where H’ = c ∘ H (pointwise scalarmultiplication) for already known vectors c (of scalars) and H (of elliptic curve points). This is more efficient than calling prove_inner_product with G and H’, but the output is a proof of the same statement. The arguments are

G_slice - a slice to the vector G of elliptic curve points

H_slice - a slice to the vector H of elliptic curve points

H_prime_scalars - a slice to the vector c of scalars such that H’ = c ∘ H

Q - the elliptic curve point Q

a_slice - a slice to the vector a of scalars

b_slice - a slice to the vector b of scalars Precondictions: G_slice, H_slice, a_slice and b_slice should all be of the same length, and this length must be a power of 2.