Function bulletproofs::range_proof::verify_efficient

pub fn verify_efficient<C: Curve>(
    transcript: &mut RandomOracle, 
    n: u8, 
    commitments: &[Commitment<C>], 
    proof: &RangeProof<C>, 
    gens: &Generators<C>, 
    v_keys: &CommitmentKey<C>
) -> Result<(), VerificationError>

This function verifies a range proof, i.e. a proof of knowledge of value v_1, v_2, …, v_m that are all in [0, 2^n) that are consistent with commitments V_i to v_i. The arguments are

n - the number n such that each v_i is claimed to be in [0, 2^n) by the prover

commitments - commitments V_i to each v_i

proof - the range proof

gens - generators containing vectors G and H both of length nm

v_keys - commitment keys B and B_tilde

This function is more efficient than the naive_verify since it unfolds what the inner product proof verifier does using the verification scalars.