Module bulletproofs::range_proof_mpc::dealer

The dealer module contains the API for the dealer state while the dealer is engaging in an aggregated multiparty computation protocol.

For more explanation of how the dealer, party, and messages modules orchestrate the protocol execution, see the API for the aggregated multiparty computation protocol.

Structs

Dealer
Used to construct a dealer for the aggregated rangeproof MPC protocol.

DealerAwaitingBitCommitments
A dealer waiting for the parties to send their BitCommitments.

DealerAwaitingPolyCommitments
A dealer which has sent the BitChallenge to the parties and is waiting for their PolyCommitments.

DealerAwaitingProofShares
A dealer which has sent the PolyChallenge to the parties and is waiting to aggregate their ProofShares into a RangeProof.