Module bulletproofs::range_proof_mpc::messages

The messages module contains the API for the messages passed between the parties and the dealer in an aggregated multiparty computation protocol.

For more explanation of how the dealer, party, and messages modules orchestrate the protocol execution, see the API for the aggregated multiparty computation protocol.

Structs

BitChallenge
Challenge values derived from all parties' BitCommitments.

BitCommitment
A commitment to the bits of a party's value.

PolyChallenge
Challenge values derived from all parties' PolyCommitments.

PolyCommitment
A commitment to a party's polynomial coefficents.

ProofShare
A party's proof share, ready for aggregation into the final RangeProof.