Module bulletproofs::inner_product_proof

Structs

InnerProductProof

VerificationScalars

This struct contains vectors of scalars that are needed for verification. Both u_sq and u_inv_sq have to be of equal length k, and s has to be of length 2^k.

Functions

inner_product

This function calculates the inner product between two vectors over any field F. The arguments are

prove_inner_product

This function computes an inner product proof, which is a proof of knowledge that the prover knows vectors a and b such that P’=<a,G>+<b,H>+<a,b>Q. This function is only used for benchmarks and testing, since prove_inner_product_with_scalars is faster when producing the inner product proof inside the range proof. On its own it could be used to produce inner product proofs, though. The arguments are

prove_inner_product_with_scalars

This function computes an inner product proof, which is a proof of knowledge that the prover knows vectors a and b such that P’=<a,G>+<b,H’>+<a,b>Q, but where H’ = c ∘ H (pointwise scalarmultiplication) for already known vectors c (of scalars) and H (of elliptic curve points). This is more efficient than calling prove_inner_product with G and H’, but the output is a proof of the same statement. The arguments are

verify_inner_product

This function verifies an inner product proof, i.e. a proof of knowledge of vectors a and b such that P’=<a,G>+<b,H>+<a,b>Q. The arguments are

verify_scalars

This function calculates the verification scalars that are used to verify an inner product proof. The arguments are